CDAP-5866 Added wrapper around AuthorizationService so that service lifecycle can be re-executed.
